move feature computation to feature subpackage
