+++ skimage/segmentation/__init__.py
@@ -13,8 +13,7 @@ from .morphsnakes import (morphological_geodesic_active_contour,
                           morphological_chan_vese, inverse_gaussian_gradient,
                           circle_level_set,
                           disk_level_set, checkerboard_level_set)
-from .trainable_segmentation import (multiscale_basic_features,
-                                     fit_segmenter, predict_segmenter,
+from .trainable_segmentation import (fit_segmenter, predict_segmenter,
                                      TrainableSegmenter)
 from ..morphology import flood, flood_fill
 
@@ -41,7 +40,6 @@ __all__ = [
     'checkerboard_level_set',
     'flood',
     'flood_fill',
-    'multiscale_basic_features',
     'fit_segmenter',
     'predict_segmenter',
     'TrainableSegmenter'
